Significant Update
Interface tidied up significantly; components now become in/visible based on the user's actions.

Various functionality added, including returning a list of all usernames. There is a bug to resolve that is stopping messages being displayed, but that functionality is 99% complete too. Messages can be sent.

Various bug fixes implemented.

Outstanding changes include implementing JPA and adding various features to do with searching messages sent/received. The interface will be further improved, but this is low priority and will be completed towards the end of the project.